perf: TaskInfo Component
